Move addBreadcrumb tests from ClientTest to HubTest
